+++ resharper/resharper-unity/test/src/CSharp/Daemon/UsageChecking/UsageInspectionsSuppressorTest.cs
@@ -19,6 +19,7 @@ namespace JetBrains.ReSharper.Plugins.Unity.Tests.CSharp.Daemon.UsageChecking
         [Test] public void SerializableClassFields01() { DoNamedTest(); }
         [Test] public void PreprocessBuildInterface01() { DoNamedTest(); }
         [Test] public void PreprocessBuildInterface02() { DoNamedTest(); }
+        [Test] public void SettingsProviderAttribute01() { DoNamedTest(); }
 
         protected override void DoTest(Lifetime lifetime, IProject project)
         {